requirements.txt
africastalking==1.2.9
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.2.1
django-allauth==0.54.0
django-extensions==4.1
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.28.0
ecdsa==0.19.1
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
oauthlib==3.2.2
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
pytest==8.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python3-openid==3.2.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9.1
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
